rename slow/ to classes/
