+++ PhysicsTools/Heppy/python/physicsobjects/Tau.py
@@ -1,1 +1,1 @@
-from slow.Tau import Tau,isTau
+from classes.Tau import Tau,isTau
