+++ PhysicsTools/Heppy/python/physicsobjects/classes/Photon.py
